Move Span.hpp to its own header
