+++ include/openPMD/RecordComponent.hpp
@@ -75,56 +75,8 @@ struct IsContiguousContainer< std::array< T_Value, N > >
 };
 } // namespace traits
 
-/**
- * @brief Subset of C++20 std::span class template.
- *
- * Any existing member behaves equivalently to those documented here:
- * https://en.cppreference.com/w/cpp/container/span
- */
 template< typename T >
-class Span
-{
-    friend class RecordComponent;
-
-private:
-    using param_t = Parameter< Operation::GET_BUFFER_VIEW >;
-    param_t m_param;
-    size_t m_size;
-    // @todo make this safe
-    Writable * m_writable;
-
-    Span( param_t param, size_t size, Writable * writable ) :
-        m_param( std::move( param ) ),
-        m_size( size ),
-        m_writable( std::move( writable ) )
-    {
-        m_param.update = true;
-    }
-
-public:
-    size_t size() const
-    {
-        return m_size;
-    }
-
-    T *data() const
-    {
-        if( m_param.out->taskSupportedByBackend )
-        {
-            // might need to update
-            m_writable->IOHandler->enqueue( IOTask( m_writable, m_param ) );
-            m_writable->IOHandler->flush();
-        }
-        return static_cast< T * >( m_param.out->ptr );
-    }
-
-    T &operator[]( size_t i ) const
-    {
-        return data()[ i ];
-    }
-};
-
-template class Span< int >;
+class Span;
 
 class RecordComponent : public BaseRecordComponent
 {
@@ -342,307 +294,6 @@ protected:
      */
     std::shared_ptr< std::string > m_name = std::make_shared< std::string >();
 }; // RecordComponent
-
-
-template< typename T >
-inline RecordComponent&
-RecordComponent::makeConstant(T value)
-{
-    if( written() )
-        throw std::runtime_error("A recordComponent can not (yet) be made constant after it has been written.");
-
-    *m_constantValue = Attribute(value);
-    *m_isConstant = true;
-    return *this;
-}
-
-template< typename T >
-inline RecordComponent&
-RecordComponent::makeEmpty( uint8_t dimensions )
-{
-    return makeEmpty( Dataset(
-        determineDatatype< T >(),
-        Extent( dimensions, 0 ) ) );
-}
-
-template< typename T >
-inline std::shared_ptr< T > RecordComponent::loadChunk(
-    Offset o, Extent e )
-{
-    uint8_t dim = getDimensionality();
-
-    // default arguments
-    //   offset = {0u}: expand to right dim {0u, 0u, ...}
-    Offset offset = o;
-    if( o.size() == 1u && o.at(0) == 0u && dim > 1u )
-        offset = Offset(dim, 0u);
-
-    //   extent = {-1u}: take full size
-    Extent extent(dim, 1u);
-    if( e.size() == 1u && e.at(0) == -1u )
-    {
-        extent = getExtent();
-        for( uint8_t i = 0u; i < dim; ++i )
-            extent[i] -= offset[i];
-    }
-    else
-        extent = e;
-
-    uint64_t numPoints = 1u;
-    for( auto const& dimensionSize : extent )
-        numPoints *= dimensionSize;
-
-    auto newData = std::shared_ptr<T>(new T[numPoints], []( T *p ){ delete [] p; });
-    loadChunk(newData, offset, extent);
-    return newData;
-}
-
-template< typename T >
-inline void RecordComponent::loadChunk(
-    std::shared_ptr< T > data,
-    Offset o,
-    Extent e )
-{
-    Datatype dtype = determineDatatype(data);
-    if( dtype != getDatatype() )
-        if( !isSameInteger< T >( dtype ) &&
-            !isSameFloatingPoint< T >( dtype ) &&
-            !isSameComplexFloatingPoint< T >( dtype ) )
-            throw std::runtime_error("Type conversion during chunk loading not yet implemented");
-
-    uint8_t dim = getDimensionality();
-
-    // default arguments
-    //   offset = {0u}: expand to right dim {0u, 0u, ...}
-    Offset offset = o;
-    if( o.size() == 1u && o.at(0) == 0u && dim > 1u )
-        offset = Offset(dim, 0u);
-
-    //   extent = {-1u}: take full size
-    Extent extent(dim, 1u);
-    if( e.size() == 1u && e.at(0) == -1u )
-    {
-        extent = getExtent();
-        for( uint8_t i = 0u; i < dim; ++i )
-            extent[i] -= offset[i];
-    }
-    else
-        extent = e;
-
-    if( extent.size() != dim || offset.size() != dim )
-    {
-        std::ostringstream oss;
-        oss << "Dimensionality of chunk ("
-            << "offset=" << offset.size() << "D, "
-            << "extent=" << extent.size() << "D) "
-            << "and record component ("
-            << int(dim) << "D) "
-            << "do not match.";
-        throw std::runtime_error(oss.str());
-    }
-    Extent dse = getExtent();
-    for( uint8_t i = 0; i < dim; ++i )
-        if( dse[i] < offset[i] + extent[i] )
-            throw std::runtime_error("Chunk does not reside inside dataset (Dimension on index " + std::to_string(i)
-                                     + ". DS: " + std::to_string(dse[i])
-                                     + " - Chunk: " + std::to_string(offset[i] + extent[i])
-                                     + ")");
-    if( !data )
-        throw std::runtime_error("Unallocated pointer passed during chunk loading.");
-
-    if( constant() )
-    {
-        uint64_t numPoints = 1u;
-        for( auto const& dimensionSize : extent )
-            numPoints *= dimensionSize;
-
-        T value = m_constantValue->get< T >();
-
-        T* raw_ptr = data.get();
-        std::fill(raw_ptr, raw_ptr + numPoints, value);
-    } else
-    {
-        Parameter< Operation::READ_DATASET > dRead;
-        dRead.offset = offset;
-        dRead.extent = extent;
-        dRead.dtype = getDatatype();
-        dRead.data = std::static_pointer_cast< void >(data);
-        m_chunks->push(IOTask(this, dRead));
-    }
-}
-
-template< typename T >
-inline void
-RecordComponent::storeChunk(std::shared_ptr<T> data, Offset o, Extent e)
-{
-    if( constant() )
-        throw std::runtime_error("Chunks cannot be written for a constant RecordComponent.");
-    if( empty() )
-        throw std::runtime_error("Chunks cannot be written for an empty RecordComponent.");
-    if( !data )
-        throw std::runtime_error("Unallocated pointer passed during chunk store.");
-    Datatype dtype = determineDatatype(data);
-    if( dtype != getDatatype() )
-    {
-        std::ostringstream oss;
-        oss << "Datatypes of chunk data ("
-            << dtype
-            << ") and record component ("
-            << getDatatype()
-            << ") do not match.";
-        throw std::runtime_error(oss.str());
-    }
-    uint8_t dim = getDimensionality();
-    if( e.size() != dim || o.size() != dim )
-    {
-        std::ostringstream oss;
-        oss << "Dimensionality of chunk ("
-            << "offset=" << o.size() << "D, "
-            << "extent=" << e.size() << "D) "
-            << "and record component ("
-            << int(dim) << "D) "
-            << "do not match.";
-        throw std::runtime_error(oss.str());
-    }
-    Extent dse = getExtent();
-    for( uint8_t i = 0; i < dim; ++i )
-        if( dse[i] < o[i] + e[i] )
-            throw std::runtime_error("Chunk does not reside inside dataset (Dimension on index " + std::to_string(i)
-                                     + ". DS: " + std::to_string(dse[i])
-                                     + " - Chunk: " + std::to_string(o[i] + e[i])
-                                     + ")");
-
-    Parameter< Operation::WRITE_DATASET > dWrite;
-    dWrite.offset = o;
-    dWrite.extent = e;
-    dWrite.dtype = dtype;
-    /* std::static_pointer_cast correctly reference-counts the pointer */
-    dWrite.data = std::static_pointer_cast< void const >(data);
-    m_chunks->push(IOTask(this, dWrite));
-}
-
-template< typename T_ContiguousContainer >
-inline typename std::enable_if<
-    traits::IsContiguousContainer< T_ContiguousContainer >::value
->::type
-RecordComponent::storeChunk(T_ContiguousContainer & data, Offset o, Extent e)
-{
-    uint8_t dim = getDimensionality();
-
-    // default arguments
-    //   offset = {0u}: expand to right dim {0u, 0u, ...}
-    Offset offset = o;
-    if( o.size() == 1u && o.at(0) == 0u && dim > 1u )
-        offset = Offset(dim, 0u);
-
-    //   extent = {-1u}: take full size
-    Extent extent(dim, 1u);
-    //   avoid outsmarting the user:
-    //   - stdlib data container implement 1D -> 1D chunk to write
-    if( e.size() == 1u && e.at(0) == -1u && dim == 1u )
-        extent.at(0) = data.size();
-    else
-        extent = e;
-
-    storeChunk(shareRaw(data), offset, extent);
-}
-
-template< typename T, typename F >
-inline Span< T >
-RecordComponent::storeChunk( Offset o, Extent e, F && createBuffer )
-{
-    if( constant() )
-        throw std::runtime_error(
-            "Chunks cannot be written for a constant RecordComponent." );
-    if( empty() )
-        throw std::runtime_error(
-            "Chunks cannot be written for an empty RecordComponent." );
-    Datatype dtype = determineDatatype<T>();
-    if( dtype != getDatatype() )
-    {
-        std::ostringstream oss;
-        oss << "Datatypes of chunk data ("
-            << dtype
-            << ") and record component ("
-            << getDatatype()
-            << ") do not match.";
-        throw std::runtime_error(oss.str());
-    }
-    uint8_t dim = getDimensionality();
-    if( e.size() != dim || o.size() != dim )
-    {
-        std::ostringstream oss;
-        oss << "Dimensionality of chunk ("
-            << "offset=" << o.size() << "D, "
-            << "extent=" << e.size() << "D) "
-            << "and record component ("
-            << int(dim) << "D) "
-            << "do not match.";
-        throw std::runtime_error(oss.str());
-    }
-    Extent dse = getExtent();
-    for( uint8_t i = 0; i < dim; ++i )
-        if( dse[i] < o[i] + e[i] )
-            throw std::runtime_error("Chunk does not reside inside dataset (Dimension on index " + std::to_string(i)
-                                     + ". DS: " + std::to_string(dse[i])
-                                     + " - Chunk: " + std::to_string(o[i] + e[i])
-                                     + ")");
-
-    /*
-     * The openPMD backend might not yet know about this dataset.
-     * Flush the openPMD hierarchy to the backend without flushing any actual
-     * data yet.
-     */
-    seriesFlush( FlushLevel::SkeletonOnly );
-
-    size_t size = 1;
-    for( auto ext : e )
-    {
-        size *= ext;
-    }
-    /*
-     * Flushing the skeleton does not create datasets,
-     * so we might need to do it now.
-     */
-    if( !written() )
-    {
-        Parameter< Operation::CREATE_DATASET > dCreate;
-        dCreate.name = *m_name;
-        dCreate.extent = getExtent();
-        dCreate.dtype = getDatatype();
-        dCreate.chunkSize = m_dataset->chunkSize;
-        dCreate.compression = m_dataset->compression;
-        dCreate.transform = m_dataset->transform;
-        dCreate.options = m_dataset->options;
-        IOHandler()->enqueue(IOTask(this, dCreate));
-    }
-    Parameter< Operation::GET_BUFFER_VIEW > getBufferView;
-    getBufferView.offset = o;
-    getBufferView.extent = e;
-    getBufferView.dtype = getDatatype();
-    IOHandler()->enqueue( IOTask( this, getBufferView ) );
-    IOHandler()->flush();
-    auto &out = *getBufferView.out;
-    if( !out.taskSupportedByBackend )
-    {
-        auto data = std::forward< F >( createBuffer )( size );
-        out.ptr = static_cast< void * >( data.get() );
-        storeChunk( std::move( data ), std::move( o ), std::move( e ) );
-    }
-    return Span< T >{ std::move( getBufferView ), size, &writable() };
-}
-
-template< typename T >
-inline Span< T >
-RecordComponent::storeChunk( Offset offset, Extent extent )
-{
-    return storeChunk< T >(
-        std::move( offset ),
-        std::move( extent ),
-        []( size_t size )
-        {
-            return std::shared_ptr< T >{
-                new T[ size ], []( auto * ptr ) { delete[] ptr; } };
-        } );
-}
 } // namespace openPMD
+
+#include "RecordComponent.tpp"
