+++ src/Mesh.cpp
@@ -230,8 +230,8 @@ Mesh::flush_impl(std::string const& name)
             for( auto& comp : *this )
             {
                 comp.second.flush(name);
-                writable()->abstractFilePosition =
-                    comp.second.writable()->abstractFilePosition;
+                writable().abstractFilePosition =
+                    comp.second.writable().abstractFilePosition;
             }
         }
         else
