+++ src/Record.cpp
@@ -79,8 +79,8 @@ Record::flush_impl(std::string const& name)
             for( auto& comp : *this )
             {
                 comp.second.flush(name);
-                writable()->abstractFilePosition =
-                    comp.second.writable()->abstractFilePosition;
+                writable().abstractFilePosition =
+                    comp.second.writable().abstractFilePosition;
             }
         }
         else
